+++ App/Screens/index.js
@@ -1,0 +1,6 @@
+
+import HomeScreen from "./Home";
+import ScannerScreen from "./Scanner";
+import SelectionScreen from "./Selection";
+
+module.exports = { HomeScreen, ScannerScreen, SelectionScreen };
